refactor(api): API 모듈 분리 및 리팩토링
- 댓글 관련 API를 `comment.ts`로 분리
- 리뷰 추천 관련 API를 `like.ts`로 분리
- `kakaoAuth.ts`를 `auth.ts`로 파일명 변경
- 알림 관련 API를 `notification.ts`로 분리
- 유저 관련 API를 `user.ts`로 분리
- 각 파일에 맞는 import 및 함수 정의 추가
- 코드 가독성을 위해 함수 및 주석 정리
