+++ front/src/api/auth.ts
@@ -21,12 +21,12 @@ export const getKakaoToken = async (code: string): Promise<void> => {
  * 카카오 서버에서 유저 정보 조회하는 함수
  * @returns 유저 정보
  */
-export const getKakaoUserInfo = async (): Promise<{
+export const getLoginUserInfo = async (): Promise<{
   isNewMember: boolean;
   userInfo: UserInfo;
 }> => {
   const response = await authApiClient.get(`/auth/kakao/user`);
-  console.log("카카오 유저 정보 조회 성공:", response.data);
+  console.log("로그인 유저 정보 조회 성공:", response.data);
 
   return response.data;
 };
